# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PRS API",
        "description": "API documentation for the PRS application.",
        "contact": {
            "email": "support@prs.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/distributor-orders/{id}": {
            "get": {
                "tags": [
                    "Distributor Orders"
                ],
                "summary": "Get details of a single distributor order",
                "operationId": "2df49f1a24a961f2774b105f9bc24e40",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "order_code": {
                                            "type": "string"
                                        },
                                        "distributor": {
                                            "type": "string"
                                        },
                                        "total_amount": {
                                            "type": "string"
                                        },
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "product_name": {
                                                        "type": "string"
                                                    },
                                                    "quantity": {
                                                        "type": "integer"
                                                    },
                                                    "batches": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "batch_no": {
                                                                    "type": "string"
                                                                },
                                                                "expiry_date": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/distributor-orders/{id}/update-status": {
            "post": {
                "tags": [
                    "Distributor Orders"
                ],
                "summary": "Update order status (Unified Status Endpoint)",
                "description": "Directly update order status. Handles logic for processing, accepted (admin), delivered (receipt), rejected, and cancelled.",
                "operationId": "5a5f2c826dba20f3d22f173698afb966",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "processing",
                                            "accepted",
                                            "delivered",
                                            "cancelled",
                                            "rejected"
                                        ]
                                    },
                                    "cancellation_reason": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}